Crate spl_program_error

Crate defining a library with a procedural macro and other dependencies for building Solana program errors

Re-exports

pub use num_derive;

pub use num_traits;

pub use solana_program;

pub use thiserror;

Attribute Macros

spl_program_error

Proc macro attribute to turn your enum into a Solana Program Error

Derive Macros

DecodeError

Derive macro to add solana_program::decode_error::DecodeError trait

IntoProgramError

Derive macro to add Into<solana_program::program_error::ProgramError> trait

PrintProgramError

Derive macro to add solana_program::program_error::PrintProgramError trait